chore: merge admin flags handler into admin-handler
Keep single admin domain file: move handleGetAdminMerchFlags /
handleSetAdminMerchFlags (with 500/400 error_code handling) from
ingest/admin-flags-handler.ts into ingest/admin-handler.ts and
delete the now-duplicate standalone file. Remove the orphaned
handleGetMerchFlags/handleSetMerchFlags exports that were never
imported (verified via grep). Update ingest/routes.ts to import
from admin-handler.js.

Verified: tsc pass, prettier pass, 854 tests pass.
